fix movie activity layout

+++ app/src/main/java/com/example/proto_korzo/database/model/Genre.java
@@ -2,20 +2,20 @@ package com.example.proto_korzo.database.model;
 
 public class Genre {
 
-    private long id;
+    private int id;
 
     private String name;
 
-    public Genre(long id, String name) {
+    public Genre(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
